Move over some sections from the introduction vignette that can be included here.  I've started moving the pieces over piecemeal because the intro vignette is too much like a reference manual and needs more of a narrative which I'm attempting to provide here.
Specifically, section 3 "Classes, Methods, and Pseudo-Methods" of the
intro vignette is 20 pages on its own of reference material that would
not be as appropriate for a writeup in R Journal or JSS or something.
It is good content, but we need to replace much of that with very
concise tables or similar.

The number of classes and methods is quite high for RProtoBuf, and so
I don't think we can describe each on in detail in this type of
writeup without losing sight of the higher level point about why this
is cool and useful.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@561 edb9625f-4e0d-4859-8d74-9fd3b1da38cb