whitespace change only.
Increase line-wrap limit to 100 characters.  This is more readable for
this code base as the large method names force too many conditionals
and assignments onto far too many lines, and it more closely matches
the previous style.

There are still a fair number of lines greater than 100 characters
that are wrapped.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@652 edb9625f-4e0d-4859-8d74-9fd3b1da38cb