Move the MapReduce / Histogram example before the OpenCPU one and greatly improve it.  Add a very simple python example of using the protoc and a few lines of python to store arrays of bin counts and breaks as a Histogram protocol buffer.
Then, use HistogramTools to read in this histogram into R, convert it
to a native R histogram object, and plot it.

Add a reference to blocker's theoretical work on preprocessing, and a
self-citation at the end of this example section to show a real
application of the design pattern described here.

TODO: All three steps - the python code, the R code, and the output
histogram plot should be much more concisely typeset onto a single
line.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@776 edb9625f-4e0d-4859-8d74-9fd3b1da38cb