Add back the article 'the' in the second sentence of the introduction, and revert back to my use of 'different' rather than various.
I believe this flows better but don't have a strong preference if
someone can explain why they like the change.

Adding the article 'the' makes it more concrete/definitive.  Certainly
others could come up with a different break down of the phases of data
analysis, but that doesn't make our breakdown wrong, and thus I see no
reason to avoid the article 'the' here, if that was indeed the reason
for the change.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@791 edb9625f-4e0d-4859-8d74-9fd3b1da38cb