Revert Jereoen's two sentences about MessagePack and BSON back to my original three sentences, and add back the citations to the relevant R packages.
Jereon's version is very dismissive of these formats as not being
widely used or compatible with the almighty JSON.

My version notes that these address real deficiencies of JSON for the
application domain we are talking about in this paper, and points out
the shortcomings they still have compared to protocol buffers.

This introduction is a natural flow of alternatives each slightly
better than the last one discussed: starting with CSV, then XML, then
JSON, then binary JSON, then protocol buffers.  For this application
domain, binary JSON is strictly better than JSON and so dismissive
comments if any should be oriented the other way towards traditional
text JSON.

The XML section Jereon and Dirk added is great, thanks.

We may still need one more sentence in the first paragraph making it
crystal clear what application domain / context is used for this
discussion of the alternatives.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@800 edb9625f-4e0d-4859-8d74-9fd3b1da38cb