Add two minor text improvements suggested by Tim Hesterberg to sections 5 and 6.
The first sentence of each of the last two paragraphs of the summary
was saying the same thing -- merge the second duplicate sentence into
the first, which then leaves the third paragraph of the summary as
just a single sentence.  Add it to the end of the second sentence and
we now have I think a slightly crisper summary in just two paragraphs.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@832 edb9625f-4e0d-4859-8d74-9fd3b1da38cb