Improve section 6 to address referee feedback:
Replace massive full page 50 row table with a more succinct plot of
the relevant data points, and label the outliers of this plot so we
can talk about the interesting cases in the text.

Add a small 3-row table beneath the plot showing the data for the two
outliers plus the aggregate of all datasets.

Explain why protocol buffers are so much more space-efficient for one
dataset, and slightly less efficient for another.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@916 edb9625f-4e0d-4859-8d74-9fd3b1da38cb